+++ src/components/PostContent/PostContent.stories.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+// also exported from '@storybook/react' if you can deal with breaking changes in 6.1
+import { Story, Meta } from "@storybook/react/types-6-0";
+
+import { PostContent as PostContentComponent, PostContentProps } from ".";
+
+export default {
+  title: "Components/Menu/Menu Item",
+  component: PostContentComponent,
+} as Meta;
+
+const Template: Story<PostContentProps> = (args) => <PostContentComponent {...args} />;
+
+export const PostContent = Template.bind({});
+PostContent.args = {};