+++ src/components/PostContent/PostContent.test.tsx
@@ -1,13 +1,13 @@
 import { render, screen } from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
 
-import { MenuItem } from ".";
+import { PostContent } from ".";
 
 test("Renders the component correctly", () => {
     // Render component
-    render(<MenuItem />)
+    render(<PostContent />)
 
     // Find the element
-    screen.getByTestId('menu-item')
+    screen.getByTestId('post-content')
 
 });