+++ src/components/Title/Title.test.tsx
@@ -1,9 +1,18 @@
-import React from "react";
-import { shallow } from "enzyme";
+import { render, screen } from '@testing-library/react'
+import '@testing-library/jest-dom/extend-expect'
+
 import { Title } from "./";
 
-test("Renders the component", () => {
-  const component = shallow(<Title label="Test" />);
-  expect(component).toMatchSnapshot();
-  expect(component.text()).toBe("Test");
+test("Renders the component correctly", () => {
+  const label = "Test Label"
+
+  // Render component
+  render(<Title label={label} />)
+
+  // Find the element
+  screen.getByTestId('title')
+
+  // Check rendered result of label prop
+  expect(screen.getByRole('heading')).toHaveTextContent(label)
+
 });